refactor: separate serializers to different files
