+++ project/users/serializers/constants.py
@@ -1,0 +1,1 @@
+VALID_SECURITY_QA_NUMS = 3
